Move coreasInterpolator to its own file for clarity
